Move team data objects to Data namespace
