+++ kits/Shared/Teams/Base/app/Concerns/HasTeams.php
@@ -2,12 +2,12 @@
 
 namespace App\Concerns;
 
+use App\Data\TeamPermissions;
+use App\Data\UserTeam;
 use App\Enums\TeamPermission;
 use App\Enums\TeamRole;
 use App\Models\Membership;
 use App\Models\Team;
-use App\Support\TeamPermissions;
-use App\Support\UserTeam;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
