+++ kits/Shared/Teams/Base/app/Data/UserTeam.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Support;
+namespace App\Data;
 
 readonly class UserTeam
 {
